+++ src/app/apps/rup/mapa-camas/views/plan-indicaciones/nueva-indicacion/nueva-indicacion.component.ts
@@ -14,7 +14,7 @@ import { PlanIndicacionesServices } from '../../../services/plan-indicaciones.se
 export class PlanIndicacionesNuevaIndicacionComponent implements OnInit {
 
     @Output() save = new EventEmitter();
-
+    @Output() edit = new EventEmitter();
     private capa: string;
     private ambito: string;
     private idInternacion: string;
@@ -25,21 +25,23 @@ export class PlanIndicacionesNuevaIndicacionComponent implements OnInit {
     registro = null;
 
     @Input() tipoPrestacion = {
-        'conceptId' : '4981000013105',
-        'term' : 'plan de indicaciones médicas',
-        'fsn' : 'plan de indicaciones médicas (procedimiento)',
-        'semanticTag' : 'procedimiento'
+        'conceptId': '4981000013105',
+        'term': 'plan de indicaciones médicas',
+        'fsn': 'plan de indicaciones médicas (procedimiento)',
+        'semanticTag': 'procedimiento'
     };
 
     @Input() seccion = null;
+    @Input() indicacion = null;
+
 
     prestacion: IPrestacion = {
         solicitud: {
             tipoPrestacion: {
-                'conceptId' : '4981000013105',
-                'term' : 'plan de indicaciones médicas',
-                'fsn' : 'plan de indicaciones médicas (procedimiento)',
-                'semanticTag' : 'procedimiento'
+                'conceptId': '4981000013105',
+                'term': 'plan de indicaciones médicas',
+                'fsn': 'plan de indicaciones médicas (procedimiento)',
+                'semanticTag': 'procedimiento'
             }
         }
     } as any;
@@ -71,13 +73,13 @@ export class PlanIndicacionesNuevaIndicacionComponent implements OnInit {
         this.selectedConcept = concepto;
         this.elementoRUP = this.elementoRUPService.buscarElemento(concepto, concepto.esSolicitud);
         this.ps.get(concepto.conceptId, true).subscribe(() => { });
-        this.registro = new IPrestacionRegistro(this.elementoRUP, concepto);
+        this.registro = this.indicacion ? { valor: this.indicacion.valor } : new IPrestacionRegistro(this.elementoRUP, concepto);
 
 
     }
 
     onSave() {
-        const nombre = this.registro.valor?.medicamento?.term || this.registro.concepto.term;
+        const nombre = this.registro.valor?.nombre || this.registro.concepto.term;
         const indicacion = {
             idInternacion: this.idInternacion,
             idPrestacion: null,
@@ -102,7 +104,24 @@ export class PlanIndicacionesNuevaIndicacionComponent implements OnInit {
                 fecha: new Date()
             }]
         };
-        this.save.emit(indicacion);
+
+        if (this.indicacion?.estadoActual.tipo === 'draft') {
+            this.indicacion.fechaInicio = indicacion.fechaInicio;
+            this.indicacion.ambito = indicacion.ambito;
+            this.indicacion.capa = indicacion.capa;
+            this.indicacion.organizacion = indicacion.organizacion;
+            this.indicacion.profesional = indicacion.profesional;
+            this.indicacion.nombre = indicacion.nombre;
+            this.indicacion.concepto = indicacion.concepto;
+            this.indicacion.valor = indicacion.valor;
+            this.indicacion.elementoRup = indicacion.elementoRUP;
+            this.indicacion.seccion = indicacion.seccion;
+            this.indicacion.estados = indicacion.estados;
+            this.edit.emit(this.indicacion);
+        } else {
+            this.save.emit(indicacion);
+        }
+
         // this.planIndicacionesServices.create(indicacion);
     }